Move internal code to internal package.
Closes #84.
