+++ radiography/api/radiography.api
@@ -3,6 +3,9 @@ public final class radiography/AttributeAppendable {
 	public final fun append (Ljava/lang/CharSequence;)V
 }
 
+public abstract interface annotation class radiography/ExperimentalRadiographyComposeApi : java/lang/annotation/Annotation {
+}
+
 public final class radiography/Radiography {
 	public static final field INSTANCE Lradiography/Radiography;
 	public static final fun scan ()Ljava/lang/String;
@@ -88,8 +91,5 @@ public final class radiography/ViewStateRenderers {
 public final class radiography/ViewsKt {
 	public static final synthetic fun scan (Landroid/view/View;Ljava/util/List;Lradiography/ViewFilter;)Ljava/lang/String;
 	public static synthetic fun scan$default (Landroid/view/View;Ljava/util/List;Lradiography/ViewFilter;ILjava/lang/Object;)Ljava/lang/String;
-}
-
-public abstract interface annotation class radiography/compose/ExperimentalRadiographyComposeApi : java/lang/annotation/Annotation {
 }
