+++ radiography/src/main/java/radiography/ExperimentalRadiographyComposeApi.kt
@@ -1,4 +1,4 @@
-package radiography.compose
+package radiography
 
 @RequiresOptIn(
     message = "This API is experimental, may only work with a specific version of Compose, " +
