+++ radiography/src/main/java/radiography/internal/Semantics.kt
@@ -1,8 +1,9 @@
-package radiography.compose
+package radiography.internal
 
 import androidx.compose.ui.semantics.SemanticsModifier
 import androidx.compose.ui.semantics.SemanticsProperties
 import radiography.ScannableView.ComposeView
+import radiography.ExperimentalRadiographyComposeApi
 
 /** Returns all tag strings set on the composable via `Modifier.testTag`. */
 @OptIn(ExperimentalRadiographyComposeApi::class)
