+++ radiography/src/main/java/radiography/internal/Strings.kt
@@ -1,4 +1,4 @@
-package radiography
+package radiography.internal
 
 internal fun CharSequence.ellipsize(maxLength: Int): CharSequence =
   if (length > maxLength) "${subSequence(0, maxLength - 1)}…" else this
