+++ radiography/src/test/java/radiography/RenderTreeStringTest.kt
@@ -2,6 +2,7 @@ package radiography
 
 import com.google.common.truth.Truth.assertThat
 import org.junit.Test
+import radiography.internal.renderTreeString
 
 class RenderTreeStringTest {
